requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apsw==3.49.1.0
apswutils==0.0.2
beautifulsoup4==4.13.3
certifi==2025.1.31
click==8.1.7
fastapi==0.109.0
fastcore==1.8.0
fastlite==0.1.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.6
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-fasthtml==0.12.5
python-multipart==0.0.20
PyYAML==6.0.1
six==1.17.0
sniffio==1.3.0
soupsieve==2.6
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.34.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
